move main area components to folder 'components'
