+++ src/app/app.component.ts
@@ -2,9 +2,9 @@ import {Component} from 'angular2/core';
 import {InitializeService} from './shared/services/initialize.service';
 import {Store} from '../store/store';
 import {LoginComponent} from './components/login/login.component';
-import {StartpageComponent} from './startpage/startpage.component';
-import {ManageComponent} from './manage/manage.component';
-import {ActivitiesComponent} from './activities/activities.component';
+import {StartpageComponent} from './components/startpage/startpage.component';
+import {ManageComponent} from './components/manage/manage.component';
+import {ActivitiesComponent} from './components/activities/activities.component';
 import {AdministrationComponent} from './administration/administration.component';
 import {OnInit} from 'angular2/core';
 import {LoginService} from './shared/stubs/services/login.service';
